+++ doc/versionHistory.rst
@@ -6,6 +6,14 @@
 Version History
 ##################
 
+.. _lsst.ts.MTAOS-0.3.5:
+
+-------------
+0.3.5
+-------------
+
+Restrict some commands can only be executed in the Enabled state.
+
 .. _lsst.ts.MTAOS-0.3.4:
 
 -------------